feat: [move-page] [kanban-task-read] 스프린트 페이지 폴더 밖으로 이동 및 디자인 수정
스프린트 페이지에서 보드 크기를 스크린 크기에 맞게 가변적으로 변할 수 있도록 수정
여러 번 사용되던 Task 타입을 sprint 타입 정의 파일로 옮김
