+++ FE/src/components/sprint/ColumnBoard.tsx
@@ -1,12 +1,12 @@
 import TaskCard from './TaskCard';
-import { Task } from '../../pages/sprint/SprintPage';
+import { Task } from '../../types/sprint';
 
 interface ColumnBoardProps {
   taskList: Task[];
 }
 
 const ColumnBoard = ({ taskList }: ColumnBoardProps) => (
-  <ul className="flex flex-col gap-3 p-5 border rounded-lg w-72 border-green-stroke">
+  <ul className="flex flex-col w-4/12 min-h-full gap-3 p-5 border rounded-lg border-green-stroke">
     {taskList.map(({ id, title, userName, point }) => (
       <li>
         <TaskCard key={id} id={id} title={title} assignee={userName} point={point} />
